Turn window configuration switcher into command
While it is possible to use `eyebrowse-next-window-config` with a
numerical prefix to switch to an arbitrary window configuration, it's
better to offer a completing interface that can be extended to display
tags as well.  I intend to offer `completing-read`,
`ido-completing-read` and helm soon to have a basic, a less basic, but
built-in with flexible matching and a fully-featured selection
interface that is able to display meta data, too.  At the moment only
`completing-read` is used which isn't too bad for an option considering
both ido and helm are able to use it for the most basic selection
tasks.

No matter what, helm will stay an "optional" dependency.