Rework modeline init
First of all, if you add a list to the modeline which has a first
element that's a normal symbol, the symbol is evaluated and the rest
only if it turns out to be truthy.  That allows one to use
`eyebrowse-mode` to display the indicator only if the mode is active, so
no more removing is necessary to stop displaying it.

The other change makes sure that nothing, be it enabling the mode a
second time non-interactively or loading a theme (see #29 for further
details) can result in a duplicated indicator.